backend/requirements.txt
fastapi
uvicorn
httpx
python-multipart
PyPDF2
google-genai
